initdb in services

+++ HaushaltServices/src/main/java/com/arash/model/ExpenseDTO.java
@@ -3,10 +3,10 @@ package com.arash.model;
 public class ExpenseDTO {
 	
 	private int userId;
-	private float sumMyExpenses;
-	private float sumEspensesOfAllUsers;
-	private float eachOneShouldPay;
-	private float myBalance;
+	private double sumMyExpenses;
+	private double sumEspensesOfAllUsers;
+	private double eachOneShouldPay;
+	private double myBalance;
 	
 	public int getUserId() {
 		return userId;
@@ -14,28 +14,28 @@ public class ExpenseDTO {
 	public void setUserId(int userId) {
 		this.userId = userId;
 	}
-	public float getSumMyExpenses() {
+	public double getSumMyExpenses() {
 		return sumMyExpenses;
 	}
-	public void setSumMyExpenses(float sumMyExpenses) {
+	public void setSumMyExpenses(double sumMyExpenses) {
 		this.sumMyExpenses = sumMyExpenses;
 	}
-	public float getSumEspensesOfAllUsers() {
+	public double getSumEspensesOfAllUsers() {
 		return sumEspensesOfAllUsers;
 	}
-	public void setSumEspensesOfAllUsers(float sumEspensesOfAllUsers) {
+	public void setSumEspensesOfAllUsers(double sumEspensesOfAllUsers) {
 		this.sumEspensesOfAllUsers = sumEspensesOfAllUsers;
 	}
-	public float getEachOneShouldPay() {
+	public double getEachOneShouldPay() {
 		return eachOneShouldPay;
 	}
-	public void setEachOneShouldPay(float eachOneShouldPay) {
+	public void setEachOneShouldPay(double eachOneShouldPay) {
 		this.eachOneShouldPay = eachOneShouldPay;
 	}
-	public float getMyBalance() {
+	public double getMyBalance() {
 		return myBalance;
 	}
-	public void setMyBalance(float myBalance) {
+	public void setMyBalance(double myBalance) {
 		this.myBalance = myBalance;
 	}
